% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayespca.R
\docType{package}
\name{bayespca}
\alias{bayespca}
\alias{bayespca-package}
\title{bayespca: Regularized Principal Component Analysis via Variational Bayes inference.}
\description{
A package for estimating PCA with a Variational Bayes algorithm.
}
\details{
`bayespca` performs Bayesian estimation of weight vectors in PCA.
 To achieve regularization, the method allows specifying fixed variances
 in the prior distributions of the weights; alternatively, it is possible
 to implement Jeffrey's and  Inverse Gamma priors on such parameters.
 In turn, the Inverse Gamma's can have fixed shape hyperparameter; and
 fixed or random scale hyperparameter. Last, the method allows performing
 component-specific Stochastic Variable Selection (`spike-and-slab` prior).
}
\section{Functions}{
 
\itemize{
    \item \code{\link{vbpca}} for model estimation; 
    \item \code{\link{vbpca_control}} for settings of control parameters; 
   \item \code{is.vbpca} for testing the class; 
    \item \code{\link{plothpdi}} for plotting high probability density intervals. 
}
}
\author{
D. Vidotto <d.vidotto@uvt.nl>
}
\keyword{Bayes,}
\keyword{PCA,}
\keyword{Search}
\keyword{Selection}
\keyword{Stochastic}
\keyword{Variable}
\keyword{Variational}
\keyword{regularized}

