% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotheatmap.R
\name{plotheatmap}
\alias{plotheatmap}
\title{Plot weights/prior precision matrix heatmaps.}
\usage{
plotheatmap(
  obj,
  matrix_type = "Tau",
  bound_tau = NULL,
  limits_heatmap = NULL,
  return_map = TRUE
)
}
\arguments{
\item{obj}{list; \cr
a \code{vbpca} object.}

\item{matrix_type}{character; \cr
a character vector defining the matrix that the user desires to plot; either 'W'
(for the weights matrix) or 'Tau' (for the prior precision matrix)}

\item{bound_tau}{float; \cr
truncating value for the prior precision matrix. This will not only cause truncation in the
Tau matrix, but also will make the weights matrix W sparse.}

\item{limits_heatmap}{array_like; \cr
2-dimensional array containing limits for the scale_fill_gradient2 functionality of ggplot2.
Meaningful only when \code{matrix_type = 'W'}.}

\item{return_map}{boolean; \cr
if TRUE, the heatmap of the matrix specified by \code{matrix_type} is plotted.}
}
\value{
the heatmap of the selected matrix, as well as the (truncated) prior precision matrix Tau, and the (sparse)
         weights matrix W.
}
\description{
Heatmap of the estimated weights (W) or prior precision matrix (Tau) estimated with \code{vbpca}.
             Furthermore, the function allows specifying a truncating (usually large) value for the prior precision
             matrix, so that all elements in the matrix exeeding such value will be truncated. This will also cause the
             corresponding elements of the weights matrix to be set to 0, making the matrix sparse. The function will
             also return such matrices.
}
\examples{

# Create a synthetic dataset
I <- 1e+3
X1 <- rnorm(I, 0, 50)
X2 <- rnorm(I, 0, 30)
X3 <- rnorm(I, 0, 10)

X <- cbind(X1, X1, X1, X2, X2, X2, X3, X3 )
X <- X + matrix(rnorm(length(X), 0, 1), ncol = ncol(X), nrow = I )
colnames(X) <- paste('X', 1:8, sep='')
# Estimate the Bayesian PCA model, with Gamma priors for tau
mod <- vbpca(X, D = 3, alphatau=1e-3, betatau=1e-3 )

# Plot heatmap of Tau (setting a threshold value at 50):
plotheatmap(mod, matrix_type='Tau', bound_tau=50)


}
\seealso{
\code{\link{vbpca}}
}
\author{
D. Vidotto <d.vidotto@uvt.nl>
}
