% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothpdi.R
\name{plothpdi}
\alias{plothpdi}
\title{Plot high posterior density intervals of vbpca objects.}
\usage{
plothpdi(obj, d = 1, X, vars = NULL)
}
\arguments{
\item{obj}{list; \cr 
a \code{vbpca} object.}

\item{d}{integer; \cr 
component number for which the intervals must be plotted.}

\item{X}{array_like; \cr 
the data matrix used to estimate the \code{vbpca} object.}

\item{vars}{array_like; \cr 
an array containing the variable numbers (column numbers) for which the intervals must be plotted. 
If nothing is specified, the method will attempt to plot intervals for all the variables of the data matrix.}
}
\value{
the graphical visualization of the high posterior density intervals.
}
\description{
Graphical visualization of the HPDI's of the weight matrix computed during the estimation of 
             \code{vbpca}. Note that no intervals will be plotted if \code{hpdi = FALSE} is specified in the 
             \code{control} argument.
}
\examples{
\dontrun{

# mod is a vbpca object estimated on the X matrix; 
# to plot the intervals of elements (1, 2, 3) of the second column of the W matrix: 
plothpdi(mod, d = 2, X, vars = 1:3)
}


}
\author{
D. Vidotto <d.vidotto@uvt.nl>
}
\seealso{
\code{\link{vbpca}}
}

