% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayespca.R
\docType{package}
\name{bayespca}
\alias{bayespca}
\title{bayespca: Regularized Principal Component Analysis via Variational Bayes inference.}
\description{
A package for estimating PCA with Variational Bayes inference.
}
\details{
Bayesian estimation of weight vectors in PCA.
 To achieve regularization, the method allows specifying fixed precisions
 in the prior distributions of the weights; alternatively, it is possible
 to specify Gamma priors for such parameters. The method allows
 for variable selection through Automatic Relevance Determination.
}
\section{Functions}{

\itemize{
    \item \code{\link{vbpca}} for model estimation;
    \item \code{is.vbpca} for testing the class;
    \item \code{\link{plotheatmap}} for plotting the precision and weights matrices;
    \item \code{\link{plothpdi}} for plotting high probability density intervals.
}
}

\references{
\itemize{

\item [1] C. M. Bishop. 'Variational PCA'. In Proc. Ninth Int. Conf. on Artificial Neural Networks.
ICANN, 1999.

}
}
\author{
D. Vidotto <d.vidotto@uvt.nl>
}
