% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothpdi.R
\name{plothpdi}
\alias{plothpdi}
\title{Plot high posterior density intervals of vbpca objects.}
\usage{
plothpdi(obj, d = 1, vars = NULL)
}
\arguments{
\item{obj}{list; \cr
a \code{vbpca} object.}

\item{d}{integer; \cr
component number for which the intervals must be plotted.}

\item{vars}{array_like; \cr
an array containing the variable numbers (column numbers) for which the intervals must be plotted.
If nothing is specified, the method will attempt to plot intervals for all the variables of the data matrix.}
}
\value{
the graphical visualization of the high posterior density intervals.
}
\description{
Graphical visualization of the HPDI's of the weight matrix computed during the estimation of
             \code{vbpca}. Note that no intervals will be plotted if \code{hpdi = FALSE} is specified in the
             \code{control} argument.
}
\examples{

# Create a synthetic dataset
I <- 1e+3
X1 <- rnorm(I, 0, 50)
X2 <- rnorm(I, 0, 30)
X3 <- rnorm(I, 0, 10)

X <- cbind(X1, X1, X1, X2, X2, X2, X3, X3 )
X <- X + matrix(rnorm(length(X), 0, 1), ncol = ncol(X), nrow = I )
colnames(X) <- paste('X', 1:8, sep='')
# Estimate the Bayesian PCA model, with Gamma priors for tau
mod <- vbpca(X, D = 3, alphatau=1e-3, betatau=1e-3, hpdi=TRUE, probHPDI=0.9 )

# Plot the intervals of variables (1, 2, 3) of the second column of the W matrix:
plothpdi(mod, d = 2, vars = 1:3)


}
\seealso{
\code{\link{vbpca}}
}
\author{
D. Vidotto <d.vidotto@uvt.nl>
}
